Add the useDisabledDocumentDragover hook
